Struct tibrv::context::Transport

pub struct Transport { /* fields omitted */ }

A struct representing a Rendezvous transport object.

A Rendezvous transport can carry messages across a network, between processes, or within a single process.

Transport is configured with three optional parameters:

The service parameter may be empty, a 'service name' or a port number. The default, or empty, value causes Rendezvous to search for a 'rendezvous' network service in NIS or /etc/services If no such name can be resolved, it defaults to a protocol specific port number.

The network parameter instructs the Rendezvous daemon to use a particular network for communications on this transport. The parameter consists of up to three parts, separated by semicolons, in the form network;multicast groups;send address. For example:

192.168.1.1 - Network interface only

eth0;255.1.1.1 - Network interface and multicast group

The daemon parameter tells the Rendezvous runtime how to find the daemon. The daemon socket is given in the format hostname:port. An empty value uses the daemon running on localhost, on the default port.

Considerable detail on the configuration and behaviour of Rendezvous transports is available in the TIBCO Rendezvous Concepts guide.

Methods

impl Transport

pub fn daemon(&self) -> Result<String, TibrvError>

Extract the daemon parameter from the transport.

pub fn network(&self) -> Result<String, TibrvError>

Extract the network parameter from the transport.

pub fn service(&self) -> Result<String, TibrvError>

Extract the service parameter from the transport.

pub fn create_inbox(&self) -> Result<String, TibrvError>

Creates an unique inbox subject name.

The created subject name will be unique for this transport.

pub fn send(&self, msg: &mut Msg) -> Result<(), TibrvError>

Send a Msg through this transport.

Note that Msg must be mutable due to the signature of the C library functions.

pub fn subscribe(&self, subject: &str) -> Result<Subscription, TibrvError>

Subscribe to a message subject.

Sets up a Rendezvous message queue, along with a callback which copies messages from the event queue into an mspc::channel for consumption from Rust.

Subject must be valid ASCII, wildcards are accepted, although a wildcard-only subject is not.

pub fn request(
&self,
msg: &mut Msg,
timeout: Option<f64>
) -> Result<Msg, TibrvError>

Send a synchronous request on the given subject, blocking until a response is received or timeout seconds have elapsed.

A None parameter for timeout means block indefinitely.

pub fn serve<F>(&self, subject: &str, f: F) -> Result<(), TibrvError> where
F: Fn(Msg) -> Result<Msg, TibrvError>,

Listen on this subject and respond to requests using the supplied closure.

The closure consumes the incoming request, and you may either re-use the same structure or create a new Msg for your response.

If reply_subject is not set on the received message (i.e it is a plain message rather than a request), it is not processed and is dropped.

pub fn async_serve<F, G>(
self,
handle: &Handle,
subject: &str,
f: F
) -> impl Future<Item = (), Error = TibrvError> where
F: Fn(Msg) -> G,
G: IntoFuture<Item = Msg, Error = TibrvError>,

pub fn async_sub(
&self,
handle: &Handle,
subject: &str
) -> Result<AsyncSub, TibrvError>

Asynchronously subscribe to a message subject.

Sets up the queue and channels as in a synchronous subscription, and returns an AsyncSub stream.

pub fn async_req(
&self,
handle: &Handle,
msg: &mut Msg
) -> Result<AsyncReq, TibrvError>

Asynchronously send a request on the given subject.

Returns an AsyncReq future representing the response.