Enum tibrv::errors::ErrorKind

pub enum ErrorKind {
    StrContentError,
    RvInitFailure,
    TransportError,
    QueueError,
    AsyncRegError,
    NonVectorFieldError,
    FieldTypeError,
    UnknownFieldTypeError(tibrv_u8),
    UnknownError(tibrv_status),
}

A list of general error categories.

This list may grow over time towards and after version 1.0, so it is not recommended to use exhaustive matches.

Variants
StrContentError

A provided string could not be converted to a CString.

RvInitFailure

The external rendezvous machinery failed to initialize.

TransportError

The rendezvous library rejected, or failed to connect to the transport.

QueueError

The producing end of an event queue closed early.

AsyncRegError

The Async callback event registration failed to complete properly.

NonVectorFieldError

A scalar MsgField was passed to a vector decoding method.

FieldTypeError

There was an attempt to decode a MsgField into a type which didn't match the internal tag.

UnknownFieldTypeError(tibrv_u8)

There was an attempt to decode a MsgField with an unsupported TIBRVMSG_* value

UnknownError(tibrv_status)

Some other Rendezvous error occurred.