Module tibrv::field

Interfaces for constructing Rendezvous Message Fields

All data to be sent via a Rendezvous message must be tagged and encapsulated in a "message field", represented here by the MsgField type.

Structs

BorrowedMsgField
A structure wrapping a MsgField

Builder
A builder for MsgField.

MsgField
A structure wrapping a tibrvMsgField

Enums

DecodedField

Traits

Decodable
Trait indicating the type may be decoded from a message field.

Encodable
Trait indicating the type may be encoded into a message field.

Functions

tibrv_encode_opaque⚠
Encode a slice as an opaque byte sequence.

tibrv_encode_port
Encode a u16 as an IP Port message field.

tibrv_try_decode_opaque⚠
Try and decode an opaque byte sequence into a slice

tibrv_try_decode_port
Try and decode an IP Port message field.