Function tibrv::field::tibrv_encode_port

pub fn tibrv_encode_port(
    port: u16, 
    name: Option<&str>, 
    id: Option<u32>
) -> MsgField

Encode a u16 as an IP Port message field.

Rendezvous has special provisions for network data types, IPv4 Addresses and IP Ports are encoded as special types using Network Byte Ordering.

Since we already provide an Impl for 'normal' u16 this function will encode using the special byte ordering.