Struct tibrv::field::Builder

pub struct Builder<'a, T: 'a> where
    T: Encodable,  { /* fields omitted */ }

A builder for MsgField.

Methods

impl<'a, T> Builder<'a, T> where
T: Encodable,

pub fn new(data: &'a T) -> Builder<'a, T>

Creates a new Builder used to construct a MsgField.o

At least one of with_name or with_id must be called to supply identifier information for the MsgField.

pub fn with_name(self, name: &'a str) -> Builder<T>

Sets the MsgField name.

pub fn with_id(self, id: u32) -> Builder<'a, T>

Sets the MsgField id.

pub fn encode(self) -> MsgField

Consumes the Builder, creating a MsgField.