Trait tibrv::field::Decodable

pub trait Decodable<'a> {
    fn tibrv_try_decode(msg: &'a MsgField) -> Result<Self, TibrvError>
    where
        Self: Sized;
}

Trait indicating the type may be decoded from a message field.

Required methods

fn tibrv_try_decode(msg: &'a MsgField) -> Result<Self, TibrvError> where
Self: Sized,

Try and decode a supplied MsgField as this type.

Rendezvous message fields include some primitive type information, but this method may fail if the sending party incorrectly encodes the data fields.

Loading content...