Trait tibrv::field::Encodable

pub trait Encodable {
    fn tibrv_encode(&self, name: Option<&str>, id: Option<u32>) -> MsgField;
}

Trait indicating the type may be encoded into a message field.

Implementations are provided for all the scalar data types supported by Rendezvous, these scalar types may in turn also be encoded as native arrays.

Also supported are strings (as &CStr), IPv4 Addresses (std::net::Ipv4Addr) and date/time, using NaiveDateTime from the chrono crate.

Used along with the Decodable trait, these methods allow seamless conversion to and from Rendezvous data structures.

Example

use tibrv::field::{Encodable, Decodable};

let array: &[u8] = &[1, 2, 3, 4];

// Encoding
let msg = array.tibrv_encode(Some("Array"), None);
assert_eq!(1, msg.inner.size);
assert_eq!(4, msg.inner.count);

// Decoding
let slice = <&[u8]>::tibrv_try_decode(&msg).unwrap();
assert_eq!([1, 2, 3, 4], slice);

Required methods

fn tibrv_encode(&self, name: Option<&str>, id: Option<u32>) -> MsgField

Encodes this variable as a message field.

Scalar types will be copied, but all vector types will be passed as pointers to the underlying C struct, so the source variable must live until the message field is added to a TibRVMsg

Arguments

At least one of name or id must be Some()

Loading content...