Struct tibrv::message::Msg

pub struct Msg { /* fields omitted */ }

A struct representing an owned Rendezvous Message.

The memory allocated to this type of Message is the responsibility of the Rust application. When this type is dropped, tibrvMsg_Destroy will be run to free any memory allocated to store the message.

Methods

impl Msg

pub fn new() -> Result<Self, TibrvError>

Construct a new owned Rendezvous Message

pub fn try_clone(&self) -> Result<Self, TibrvError>

pub fn add_field(
&mut self,
field: &mut MsgField
) -> Result<&mut Self, TibrvError>

Add a MsgField to this message.

The referenced field must be marked mutable, as although the process should not mutate the field, the C library makes no guarantee.

The contents of message fields are always copied, therefore slice types must be Copy. A borrowed MsgField does not need to live beyond the point where it is added to the Msg.

pub fn get_field_by_name<'a>(
&'a self,
name: &str
) -> Result<BorrowedMsgField<'a>, TibrvError>

Get a specified field from this message.

Data in scalar fields is copied, and data in pointer fields is guaranteed to live at least as long as the parent Msg.

This variant retrieves the field by name.

pub fn get_field_by_id(&self, id: u32) -> Result<BorrowedMsgField, TibrvError>

Get a specified field from this message.

Data in scalar fields is copied, and data in pointer fields is guaranteed to live at least as long as the parent Msg.

This variant retrieves the field by id.

pub fn get_field_by_index(
&self,
index: u32
) -> Result<BorrowedMsgField, TibrvError>

Get a specified field from this message.

Data in scalar fields is copied, and data in pointer fields is guaranteed to live at least as long as the parent Msg.

This variant retrieves the field by index.

pub fn remove_field_by_name(&self, name: &str) -> Result<(), TibrvError>

Remove a specified field from this message.

Data in scalar fields is copied, and data in pointer fields is guaranteed to live at least as long as the parent Msg.

This variant retrieves the field by name.

pub fn remove_field_by_id(&self, id: u32) -> Result<(), TibrvError>

Remove a specified field from this message.

Data in scalar fields is copied, and data in pointer fields is guaranteed to live at least as long as the parent Msg.

This variant retrieves the field by id.

pub fn num_fields(&self) -> Result<u32, TibrvError>

Get the number of fields within this message.

pub fn expand(&mut self, amount: i32) -> Result<&mut Self, TibrvError>

Expand the internal storage of a message.

Messages automatically expand when adding a field would overflow the available space, however if adding a large number of fields it may be useful to preallocate enough space to hold them all.

pub fn byte_size(&self) -> Result<u32, TibrvError>

Get the size of the message (in bytes).

Does not include space allocated but not yet used.

pub fn get_send_subject(&self) -> Result<Option<String>, TibrvError>

Gets the send subject for the message.

Will return Ok(None) if the send subject is empty.

pub fn set_send_subject(&mut self, subject: &str) -> Result<(), TibrvError>

Set the send subject for the message.

No wildcards are permitted in sender subjects.

pub fn get_reply_subject(&self) -> Result<Option<String>, TibrvError>

Gets the reply subject for the message.

Will return Ok(None) if the reply subject is empty.

pub fn set_reply_subject(&mut self, subject: &str) -> Result<(), TibrvError>

Sets the reply subject for the message.

No wildcards are permitted in reply subjects.